url_monitor/requirements.txt
confluent-kafka
docopt
psycopg2-binary
PyYAML
schema
